refactoring Aircraft class
