+++ pom.xml
@@ -40,7 +40,6 @@
             <artifactId>poi-ooxml</artifactId>
             <version>5.0.0</version>
         </dependency>
-
     </dependencies>
 
     <properties>
